+++ src/libs/context/contextLogin.tsx
@@ -1,19 +1,6 @@
-import { ReactNode, createContext, useContext, useState } from 'react';
+import { ReactNode, useState } from 'react';
 
-type AuthContextType = {
-  isLogged: boolean;
-  setIsLogged: (value: boolean) => void;
-};
-
-const AuthContext = createContext<AuthContextType | null>(null);
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error('useAuth는 LoginProvider 안에서만 사용해야 합니다.');
-  }
-  return context;
-};
+import { AuthContext } from '../hooks';
 
 export const LoginProvider = ({ children }: { children: ReactNode }) => {
   const [isLogged, setIsLogged] = useState(false);
